Refactor: Move CurveColor.cs, CurveFloat.cs, CurveInt.cs, and CurveVector2.cs to seperate files within Core.Structs namespace. (All curve class will be moved to Core namespace later.)
Refactor: Curve.cs, CurveColors.cs, CurveFloat.cs, CurveInt.cs, and CurveVector2.cs todo added for moving to different namespace.

Refactor: BitFlag.cs cleanup.
