+++ ShapeEngine/Core/Structs/BitFlag.cs
@@ -16,13 +16,6 @@ public readonly struct BitFlag : IBitwiseOperators<BitFlag, BitFlag, BitFlag>, I
             Flag |= v;
         }
     }
-    // public BitFlag(params int[] values)
-    // {
-    //     foreach (var v in values)
-    //     {
-    //         Flag |= (uint)v;
-    //     }
-    // }
     #region Operations Uint
     public static BitFlag Clear() => Empty;
     public bool IsEmpty() => Flag == 0;
@@ -221,97 +214,3 @@ public readonly struct BitFlag : IBitwiseOperators<BitFlag, BitFlag, BitFlag>, I
     public static BitFlag operator ^(uint left, BitFlag right) => new(left ^ right.Flag);
     #endregion
 }
-
-
-// public static class ShapeFlag
-// {
-//     #region Int
-//
-//     public static int GetFlagInt(int power) => (int)MathF.Pow(2, power);
-//     public static int ClearIntFlag()
-//     {
-//         return 0;
-//     }
-//     public static bool HasIntFlag(this int flag, int value)
-//     {
-//         if (flag == 0) return false;
-//         
-//         return (flag & value) != 0;
-//     }
-//     public static int SetIntFlag(this int flag, int value)
-//     {
-//         return flag ^= value;
-//     }
-//     public static int AddIntFlag(this int flag, int value)
-//     {
-//         return flag |= value;
-//     }
-//     public static int RemoveIntFlag(this int flag, int value)
-//     {
-//         return flag &= ~value;
-//     }
-//     public static int AddIntFlag(this int flag, params int[] values)
-//     {
-//         foreach (var v in values)
-//         {
-//             flag |= v;
-//         }
-//
-//         return flag;
-//     }
-//     public static int RemoveIntFlag(this int flag, params int[] values)
-//     {
-//         foreach (var v in values)
-//         {
-//             flag &= ~v;
-//         }
-//
-//         return flag;
-//     }
-//     #endregion
-//     
-//     #region Uint
-//     public static uint GetFlagUint(uint power) => (uint)MathF.Pow(2, power);
-//     public static uint ClearUintFlag()
-//     {
-//         return 0;
-//     }
-//     public static bool HasUintFlag(this uint flag, uint value)
-//     {
-//         if (flag == 0) return false;
-//         
-//         return (flag & value) != 0;
-//     }
-//     public static uint SetUintFlag(this uint flag, uint value)
-//     {
-//         return flag ^= value;
-//     }
-//     public static uint AddUintFlag(this uint flag, uint value)
-//     {
-//         return flag |= value;
-//     }
-//     public static uint RemoveUintFlag(this uint flag, uint value)
-//     {
-//         return flag &= ~value;
-//     }
-//     public static uint AddUintFlag(this uint flag, params uint[] values)
-//     {
-//         foreach (var v in values)
-//         {
-//             flag |= v;
-//         }
-//
-//         return flag;
-//     }
-//     public static uint RemoveUintFlag(this uint flag, params uint[] values)
-//     {
-//         foreach (var v in values)
-//         {
-//             flag &= ~v;
-//         }
-//
-//         return flag;
-//     }
-//     #endregion
-//     
-// }
